+++ examples/dashboard/src/main.tsx
@@ -2,9 +2,12 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./components/App";
 import "./css/daisyui.css";
+// import { MembraneContextProvider } from "./jellifishClientSetup";
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    {/*<MembraneContextProvider>*/}
+      <App />
+    {/*</MembraneContextProvider>*/}
   </React.StrictMode>
 );
